+++ pyogrio/__init__.py
@@ -3,6 +3,7 @@ from pyogrio.core import (
     read_info,
     set_gdal_config_options,
     get_gdal_config_option,
+    __gdal_version__,
 )
 from pyogrio.geopandas import read_dataframe, write_dataframe
 from pyogrio._version import get_versions